+++ packages/gatsby/src/bootstrap/get-config-file.js
@@ -3,7 +3,6 @@ const levenshtein = require(`fast-levenshtein`)
 const fs = require(`fs-extra`)
 const testRequireError = require(`../utils/test-require-error`).default
 const report = require(`gatsby-cli/lib/reporter`)
-const chalk = require(`chalk`)
 const path = require(`path`)
 const existsSync = require(`fs-exists-cached`).sync
 
@@ -32,20 +31,30 @@ module.exports = async function getConfigFile(
       })
     )
     if (!testRequireError(configPath, err)) {
-      report.panic(
-        `We encountered an error while trying to load your site's ${configName}. Please fix the error and try again.`,
-        err
-      )
+      report.panic({
+        id: `10123`,
+        error: err,
+        context: {
+          configName,
+          message: err.message,
+        },
+      })
     } else if (nearMatch) {
-      report.panic(
-        `It looks like you were trying to add the config file? Please rename "${chalk.bold(
-          nearMatch
-        )}" to "${chalk.bold(configName)}"`
-      )
-    } else if (existsSync(path.join(rootDir, `src`, configName))) {
-      report.panic(
-        `Your ${configName} file is in the wrong place. You've placed it in the src/ directory. It must instead be at the root of your site next to your package.json file.`
-      )
+      report.panic({
+        id: `10124`,
+        error: err,
+        context: {
+          configName,
+          nearMatch,
+        },
+      })
+    } else if (existsSync(path.join(rootDir, `src`, configName + `.js`))) {
+      report.panic({
+        id: `10125`,
+        context: {
+          configName,
+        },
+      })
     }
   }