+++ packages/gatsby-plugin-google-analytics/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-analytics/compare/gatsby-plugin-google-analytics@2.1.0...gatsby-plugin-google-analytics@2.1.1) (2019-06-27)
+
+**Note:** Version bump only for package gatsby-plugin-google-analytics
+
 # [2.1.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-analytics/compare/gatsby-plugin-google-analytics@2.0.21...gatsby-plugin-google-analytics@2.1.0) (2019-06-20)
 
 **Note:** Version bump only for package gatsby-plugin-google-analytics