+++ packages/gatsby-source-shopify/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-shopify/compare/gatsby-source-shopify@2.1.2...gatsby-source-shopify@2.1.3) (2019-07-02)
+
+**Note:** Version bump only for package gatsby-source-shopify
+
 ## [2.1.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-shopify/compare/gatsby-source-shopify@2.1.1...gatsby-source-shopify@2.1.2) (2019-06-25)
 
 **Note:** Version bump only for package gatsby-source-shopify